proper gtest macros, additional executable to regenerate iir of in-memory stencils
